+++ start/routes.ts
@@ -72,7 +72,6 @@ Route.get('/usuarios', 'UsuariosController.index')
 Route.get('/usuarios/:email', 'UsuariosController.indexByEmail')
 Route.get('/usuarios_receitas/:id', 'UsuariosController.indexById')
 Route.post('/usuarios', 'UsuariosController.store')
-Route.post('/alterar_senha/:email', 'UsuariosController.AlteracaoDeSenha')
 Route.put('/usuarios/:id', 'UsuariosController.update')
 Route.delete('/usuarios/:id', 'UsuariosController.destroy')
 
@@ -138,3 +137,4 @@ Route.delete('/exame_marcados/:id', 'ExamesMarcadosController.destroy')
 
 Route.post('/login', 'SessoesController.login')
 Route.post('/verificar', 'SessoesController.verificarSenha')
+Route.post('/alterar_senha/:email', 'SessoesController.AlteracaoDeSenha')
